\name{conc.data3}
\alias{conc.data3}
\docType{data}
\title{
Dataset with faked concentration records of Adda River}
\description{
Scattered concentration data of Total Nitrogen ("TN") in mg/L. The data were made up for explanatory purpouse. The dataset covers a one-year period, from January 2017 to December 2017.}
\usage{data("conc.data3")}
\format{
  A data frame with 43 observations on the following 2 variables.
  \describe{
    \item{\code{datetime}}{a POSIXct}
    \item{\code{TN}}{a numeric vector}
  }
}
\examples{
data("conc.data3")
str(conc.data3)
}
\keyword{datasets}
